+++ pkg/engagement/presentation/graph/mailgun.graphql
@@ -1,4 +1,3 @@
 extend type Mutation {
-  simpleEmail(subject: String!, text: String!, to: [String!]!): String!
   testFeature: Boolean!
 }
