+++ pkg/engagement/presentation/graph/resolver.go
@@ -15,7 +15,7 @@ import (
 //
 // It serves as dependency injection for your app, add any dependencies you require here.
 
-//go:generate go run github.com/99designs/gqlgen
+//go:generate go run github.com/savannahghi/engagement-service/cmd/generator
 
 // Resolver sets up a GraphQL resolver with all necessary dependencies
 type Resolver struct {
