+++ pkg/engagement/presentation/imported/mailgun.graphql
@@ -1,0 +1,3 @@
+extend type Mutation {
+  simpleEmail(subject: String!, text: String!, to: [String!]!): String!
+}
