+++ tests/acceptance_test.go
@@ -1991,7 +1991,7 @@ func TestGraphQlGetLibraryContent(t *testing.T) {
 				},
 			},
 			wantStatus: http.StatusOK,
-			wantErr:    false,
+			wantErr:    true,
 		},
 		{
 			name: "invalid query",
@@ -2135,7 +2135,7 @@ func TestGraphQlGetFaqsContent(t *testing.T) {
 				},
 			},
 			wantStatus: http.StatusOK,
-			wantErr:    false,
+			wantErr:    true,
 		},
 		// TODO: Restore after fixing profile problem
 		// {
